Add files via upload
This code defines a Flutter application that has a custom HttpClient to handle bad certificates.

The MyHttpOverrides class extends the HttpOverrides class and overrides the createHttpClient method. This method returns an HttpClient object with a custom badCertificateCallback that always returns true, which allows the client to ignore bad certificates.

The main function sets the global HttpOverrides to an instance of MyHttpOverrides and then runs the Flutter app.

The Flutter app itself is a MaterialApp with a custom theme and a home page defined as the TabsPage widget. The app also uses the MultiProvider widget to provide a NewsService to the app's descendants. The NewsService is created using the ChangeNotifierProvider widget.