Add files via upload
This code defines a Flutter app that has two screens: HomeScreen and LoginScreen. It sets the app's initial route to be login, which means that the LoginScreen will be displayed when the app is launched. It also defines two routes in the app's routes map, which allows you to navigate to the other screens by calling Navigator.pushNamed with the corresponding route name.

The app's theme is set to ThemeData.light() with the scaffoldBackgroundColor property set to Colors.grey[300], which means that it will use a light theme with a grey background color for the scaffold. The app also disables the debug banner.

The main() function calls runApp with the MyApp widget as an argument, which runs the app and displays it on the screen